requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.4
gunicorn==22.0.0
coverage==7.8.0
